Syncing with experimental controller

+++ sar_control/include/nml.h
@@ -37,7 +37,7 @@ extern "C" {
 typedef struct nml_mat_s {
   int num_rows;
   int num_cols;
-  double **data;
+  float **data;
   int is_square;
 } nml_mat;
 
@@ -60,18 +60,18 @@ typedef struct nml_mat_qr_s {
 // *****************************************************************************
 
 nml_mat *nml_mat_new(int num_rows, int num_cols);
-nml_mat *nml_mat_rnd(int num_rows, int num_cols, double min, double max);
+nml_mat *nml_mat_rnd(int num_rows, int num_cols, float min, float max);
 nml_mat *nml_mat_sqr(int size);
-nml_mat *nml_mat_sqr_rnd(int size, double min, double max);
+nml_mat *nml_mat_sqr_rnd(int size, float min, float max);
 nml_mat *nml_mat_eye(int size);
 nml_mat *nml_mat_cp(nml_mat *m);
 int nml_mat_cp_r(nml_mat *m, nml_mat *m_cp);
-nml_mat *nml_mat_from(int num_rows, int num_cols, int n_vals, double *vals);
+nml_mat *nml_mat_from(int num_rows, int num_cols, int n_vals, float *vals);
 nml_mat *nml_mat_fromfile(const char *file);
 nml_mat *nml_mat_fromfilef(FILE *f);
 nml_mat* nml_mat_fromstr(char* str);
 void nml_mat_free(nml_mat *matrix);
-void nml_mat_fill_fromarr(nml_mat *m,int num_rows, int num_cols, int n_vals, double *vals);
+void nml_mat_fill_fromarr(nml_mat *m,int num_rows, int num_cols, int n_vals, float *vals);
 
 // *****************************************************************************
 //
@@ -79,7 +79,7 @@ void nml_mat_fill_fromarr(nml_mat *m,int num_rows, int num_cols, int n_vals, dou
 //
 // *****************************************************************************
 int nml_mat_eqdim(nml_mat *m1, nml_mat *m2);
-int nml_mat_eq(nml_mat *m1, nml_mat *m2, double tolerance);
+int nml_mat_eq(nml_mat *m1, nml_mat *m2, float tolerance);
 
 // *****************************************************************************
 //
@@ -94,20 +94,20 @@ void nml_mat_printf(nml_mat *matrix, const char *d_fmt);
 // Accessing and modifying matrix elements
 //
 // *****************************************************************************
-double nml_mat_get(nml_mat *matrix, int i, int j);
-void nml_mat_set(nml_mat *matrix, int i, int j, double value);
+float nml_mat_get(nml_mat *matrix, int i, int j);
+void nml_mat_set(nml_mat *matrix, int i, int j, float value);
 nml_mat *nml_mat_col_get(nml_mat *m, int col);
-nml_mat *nml_mat_col_mult(nml_mat *m, int col, double num);
-int nml_mat_col_mult_r(nml_mat *m, int col, double num);
+nml_mat *nml_mat_col_mult(nml_mat *m, int col, float num);
+int nml_mat_col_mult_r(nml_mat *m, int col, float num);
 nml_mat *nml_mat_row_get(nml_mat *m, int row);
-nml_mat *nml_mat_row_mult(nml_mat *m, int row, double num);
-int nml_mat_row_mult_r(nml_mat *m, int row, double num);
-void nml_mat_all_set(nml_mat *matrix, double value);
-int nml_mat_diag_set(nml_mat *matrix, double value);
-nml_mat *nml_mat_row_addrow(nml_mat *m, int where, int row, double multiplier);
-int nml_mat_row_addrow_r(nml_mat *m, int where, int row, double multiplier);
-nml_mat *nml_mat_smult(nml_mat *m, double num);
-int nml_mat_smult_r(nml_mat *m, double num);
+nml_mat *nml_mat_row_mult(nml_mat *m, int row, float num);
+int nml_mat_row_mult_r(nml_mat *m, int row, float num);
+void nml_mat_all_set(nml_mat *matrix, float value);
+int nml_mat_diag_set(nml_mat *matrix, float value);
+nml_mat *nml_mat_row_addrow(nml_mat *m, int where, int row, float multiplier);
+int nml_mat_row_addrow_r(nml_mat *m, int where, int row, float multiplier);
+nml_mat *nml_mat_smult(nml_mat *m, float num);
+int nml_mat_smult_r(nml_mat *m, float num);
 
 // *****************************************************************************
 //
@@ -134,7 +134,7 @@ nml_mat *nml_mat_sub(nml_mat *m1, nml_mat *m2);
 int nml_mat_sub_r(nml_mat *m1, nml_mat *m2);
 nml_mat *nml_mat_dot(nml_mat *m1, nml_mat *m2);
 nml_mat *nml_mat_transp(nml_mat *m);
-double nml_mat_trace(nml_mat* m);
+float nml_mat_trace(nml_mat* m);
 
 
 
@@ -157,7 +157,7 @@ nml_mat_lup *nml_mat_lup_solve(nml_mat *m);
 void nml_mat_lup_free(nml_mat_lup* lu);
 void nml_mat_lup_print(nml_mat_lup *lu);
 void nml_mat_lup_printf(nml_mat_lup *lu, const char *fmt);
-double nml_mat_det(nml_mat_lup* lup);
+float nml_mat_det(nml_mat_lup* lup);
 nml_mat *nml_mat_lu_get(nml_mat_lup* lup);
 nml_mat *nml_mat_inv(nml_mat_lup *m);
 
@@ -177,9 +177,9 @@ nml_mat *nml_ls_solve(nml_mat_lup *lup, nml_mat* b);
 //
 // *****************************************************************************
 
-double nml_vect_dot(nml_mat *m1, int m1col, nml_mat *m2, int m2col);
+float nml_vect_dot(nml_mat *m1, int m1col, nml_mat *m2, int m2col);
 nml_mat *nml_mat_l2norm(nml_mat *m);
-double nml_mat_col_l2norm(nml_mat *m1, int j);
+float nml_mat_col_l2norm(nml_mat *m1, int j);
 nml_mat *nml_mat_normalize(nml_mat *m);
 int nml_mat_normalize_r(nml_mat *m);
 nml_mat_qr *nml_mat_qr_new();
@@ -193,10 +193,10 @@ nml_mat_qr * nml_mat_qr_solve(nml_mat *m);
 //
 // *****************************************************************************
 nml_mat *nml_mat_funcElement(nml_mat *m,float (*Function)(float x));
-double nml_mat_sum_elem(nml_mat *m);
+float nml_mat_sum_elem(nml_mat *m);
 nml_mat* extend_row_vec(nml_mat* vec, int num_rows);
 void nml_mat_print_CF(nml_mat *matrix);
-double str_to_double(const char *str);
+float str_to_double(const char *str);
 
 
 #ifdef __cplusplus
